+++ src/App.jsx
@@ -1,6 +1,6 @@
 import { RouterProvider } from "react-router-dom";
 import routes from './Config/Routes';
-import { AppStateProvider } from './store/provider';
+import { AppStateProvider } from './Config/store/provider';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import './App.css'
