+++ src/Config/Routes.jsx
@@ -2,8 +2,8 @@ import { createBrowserRouter } from "react-router-dom";
 import Dashboard from "../Home/Dashboard.component";
 import Manufacturers from "../Admin/Manufacturer/Manufacturers.component";
 import GenresContainer from "../Admin/Genre/Genres.container";
-import Layout from '../HOC/Layout/Layout.component';
-import ErrorBoundaryLayout from "../HOC/Layout/ErrorBoundaryLayout";
+import Layout from '../Layout/Layout.component';
+import ErrorBoundaryLayout from "../Layout/ErrorBoundaryLayout";
 
 
 const router = createBrowserRouter([
