+++ src/Layout/ErrorBoundaryLayout.jsx
@@ -1,5 +1,5 @@
 import { Outlet } from "react-router-dom";
-import { ErrorBoundary } from "../../Config/ErrorBoundary";
+import { ErrorBoundary } from "../Config/ErrorBoundary";
 
 const ErrorBoundaryLayout = () => (
   <ErrorBoundary>
